+++ src/components/MovieItem.ts
@@ -1,6 +1,6 @@
 import star_filled from '../assets/star_filled.png';
 
-import { IMovie } from '../data/api';
+import { IMovie } from '../api';
 import { makeImagePath } from '../utils/makeImagePath';
 import { fixDecimal } from '../utils/fixDecimal';
